feat: PendingDonations
